# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/feature-requests:
    get:
      operationId: adminListFeatureRequests
      tags: [admin]
      summary: Admin — list all feature requests across all users
      responses:
        "200":
          description: All feature requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureRequest"
        "403":
          description: Forbidden
  /admin/ai-prioritization:
    get:
      operationId: adminGetAiPrioritization
      tags: [admin]
      summary: Admin — get the latest AI-generated priority ranking of "requested" feature requests
      responses:
        "200":
          description: Ranked list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiPrioritization"
        "403":
          description: Forbidden
  /admin/ai-prioritization/refresh:
    post:
      operationId: adminRefreshAiPrioritization
      tags: [admin]
      summary: Admin — start a new AI priority ranking run in the background
      description: Returns immediately with the current snapshot (isRunning=true). Poll GET /admin/ai-prioritization until isRunning flips to false and generatedAt has advanced.
      responses:
        "202":
          description: Run started in the background
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiPrioritization"
        "403":
          description: Forbidden
        "409":
          description: A run is already in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiPrioritization"
  /admin/conversations/{id}:
    get:
      operationId: adminGetConversation
      tags: [admin]
      summary: Admin — get any conversation with full transcript
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "403":
          description: Forbidden
        "404":
          description: Not found
  /admin/conversations/{id}/messages:
    post:
      operationId: adminPostMessage
      tags: [admin]
      summary: Admin — post a clarifying question into a user's conversation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminMessageInput"
      responses:
        "200":
          description: Updated conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /admin/stats:
    get:
      operationId: adminStats
      tags: [admin]
      summary: Admin — aggregate stats across all feature requests
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403":
          description: Forbidden
...
components:
  schemas:
...
    Message:
      type: object
      required: [id, role, content, createdAt, attachments]
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant, admin]
        authorName:
          type: ["string", "null"]
          description: For admin messages, the name/email of the admin who posted
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
    Attachment:
      type: object
      required: [id, filename, mimeType, sizeBytes, createdAt]
      properties:
        id:
          type: integer
        filename:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        createdAt:
          type: string
          format: date-time
    Conversation:
      type: object
      required: [id, title, status, minor, createdAt, updatedAt, messages, pendingAttachments]
      properties:
        id:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [active, finalized]
        minor:
          type: boolean
          description: Filed via the "Minor" quick-report path.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        featureRequestId:
          type: ["integer", "null"]
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        pendingAttachments:
          type: array
          description: Attachments uploaded but not yet attached to a sent message
          items:
            $ref: "#/components/schemas/Attachment"
    AdminMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
...
    FeatureRequest:
      type: object
      required:
        - id
        - userId
        - userEmail
        - title
        - summary
        - problem
        - benefits
        - currentSpend
        - scope
        - status
        - priority
        - createdAt
        - updatedAt
        - conversationId
        - minor
      properties:
        id:
          type: integer
        userId:
          type: string
        userEmail:
          type: string
        userName:
          type: ["string", "null"]
        title:
          type: string
        summary:
          type: string
        problem:
          type: string
          description: Why this is a problem for the customer
        benefits:
          type: string
          description: Benefits/value the feature delivers
        currentSpend:
          type: string
          description: Money/time the customer spends today on this problem
        scope:
          type: string
          description: Full scoped requirements as Markdown
        status:
          type: string
          enum: [requested, ready_for_execution, planned, in_progress, deployed]
        priority:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        conversationId:
          type: integer
        minor:
          type: boolean
          description: Filed via the "Minor" quick-report path (small bug/feature).
        aiPriorityRank:
          type: ["integer", "null"]
          description: 1-based rank assigned by the periodic AI prioritization job (only set for "requested" items in the most recent run).
        aiPriorityRationale:
          type: ["string", "null"]
          description: Short explanation of why this request received its AI rank.
        aiPriorityGeneratedAt:
          type: ["string", "null"]
          format: date-time
          description: When the current AI rank was generated.
        relatedRequestIds:
          type: ["array", "null"]
          description: Admin-only. Ids of other feature requests this looks like a near-duplicate of or strongly clusters with. Surface only in admin UIs.
          items:
            type: integer
        clusterRationale:
          type: ["string", "null"]
          description: Admin-only. Short explanation of how this request relates to the items in relatedRequestIds. Never shown to the requesting customer.
        adminPriorityRank:
          type: ["integer", "null"]
          description: Admin-only. Manual 1-based rank set by an admin when moving a request to "Planned". Lower = higher priority. Used to order the backlog fed to the external dev agent. Defaults to the AI rank but can be overridden.
        paperclipIssueId:
          type: ["string", "null"]
          description: Admin-only. ID of the Paperclip issue created for this request, or null if not yet pushed.
        paperclipIdentifier:
          type: ["string", "null"]
          description: Admin-only. Human-readable Paperclip identifier (e.g. "PUL-23").
        paperclipStatus:
          type: ["string", "null"]
          description: Admin-only. Latest Paperclip status (open/in_progress/in_review/done/cancelled/...).
        paperclipPushedAt:
          type: ["string", "null"]
          format: date-time
          description: Admin-only. When this request was first pushed to Paperclip.
        paperclipLastSyncedAt:
          type: ["string", "null"]
          format: date-time
          description: Admin-only. When the Paperclip snapshot was last refreshed by the poller.
        paperclipPushError:
          type: ["string", "null"]
          description: Admin-only. Last push error message, if the most recent push attempt failed.
        engineeringOwner:
          type: ["string", "null"]
          enum: [agent, human, null]
          description: Admin-only. Engineering destination chosen at plan-time. "agent" routes to Paperclip + the dev-agent queue; "human" routes to Notion only. null = not yet routed (legacy rows are treated as agent-eligible).
        notionPageId:
          type: ["string", "null"]
          description: Admin-only. ID of the Notion page created for this human-routed request, or null if not yet pushed.
        notionUrl:
          type: ["string", "null"]
          description: Admin-only. Direct URL to the Notion page.
        notionStatus:
          type: ["string", "null"]
          description: Admin-only. Latest Notion Status value (e.g. Unpicked / In progress / Done).
        notionAssignee:
          type: ["string", "null"]
          description: Admin-only. Display name of the human assignee synced from Notion, if any.
        notionPushedAt:
          type: ["string", "null"]
          format: date-time
          description: Admin-only. When this request was first pushed to Notion.
        notionLastSyncedAt:
          type: ["string", "null"]
          format: date-time
          description: Admin-only. When the Notion snapshot was last refreshed by the poller.
        notionPushError:
          type: ["string", "null"]
          description: Admin-only. Last Notion push error message, if the most recent push attempt failed.
...
    AdminStats:
      type: object
      required: [total, byStatus, byPriority, recentlyCreated]
      properties:
        total:
          type: integer
        byStatus:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"
        byPriority:
          type: array
          items:
            $ref: "#/components/schemas/PriorityCount"
        recentlyCreated:
          type: array
          items:
            $ref: "#/components/schemas/FeatureRequest"
    AiPrioritization:
      type: object
      required: [items, isRunning]
      properties:
        generatedAt:
          type: ["string", "null"]
          format: date-time
          description: When the most recent ranking was generated; null if it has never run.
        nextRunAt:
          type: ["string", "null"]
          format: date-time
          description: When the next scheduled auto-run will fire.
        isRunning:
          type: boolean
          description: True while a ranking run is in progress. Clients should poll until this flips to false.
        items:
          type: array
          items:
            $ref: "#/components/schemas/AiPriorityItem"
    AiPriorityItem:
      type: object
      required:
        - featureRequestId
        - rank
        - rationale
        - title
        - summary
        - status
        - priority
        - userEmail
        - createdAt
      properties:
        featureRequestId:
          type: integer
        rank:
          type: integer
        rationale:
          type: string
        title:
          type: string
        summary:
          type: string
        status:
          type: string
          enum: [requested, ready_for_execution, planned, in_progress, deployed]
        priority:
          type: string
          enum: [low, medium, high]
        userEmail:
          type: string
        userName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    StatusCount:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: integer
    PriorityCount:
      type: object
      required: [priority, count]
      properties:
        priority:
          type: string
        count:
          type: integer
...